requirements.txt
Django==5.0.3
djangorestframework==3.15.0
whitenoise==6.6.0
python-dotenv==0.21.0
tensorflow==2.15.0
Pillow==10.0.1
numpy==1.24.3